# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/phieu-muon": {
      "post": {
        "description": "Trong một transaction: `sp_tao_phieu_muon` rồi `sp_them_sach_vao_phieu` cho từng bản sách. Một cuốn bị từ chối (đã mượn tối đa số sách cho phép, bản không sẵn sàng, người dùng không đủ điều kiện mượn...) thì không để lại phiếu nào và trả 422 kèm lý do. Cán bộ lập phiếu lấy từ JWT.\n\n**Quyền:** `ADMIN`, `THU_THU`",
        "operationId": "MuonTra_taoPhieu",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaoPhieuMuonDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhieuMuonChiTietDto"
                }
              }
            }
          },
          "400": {
            "description": "Dữ liệu gửi lên không hợp lệ (message là mảng lỗi validate) hoặc vi phạm CHECK của CSDL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Thiếu/sai/hết hạn JWT, hoặc tài khoản/người dùng đang bị khóa hay ngừng hoạt động",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Vai trò của tài khoản không được phép gọi endpoint này",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy bản ghi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "422": {
            "description": "Vi phạm quy tắc nghiệp vụ do procedure/trigger của CSDL báo (message là thông báo của CSDL)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lập phiếu mượn",
        "tags": [
          "muon-tra"
        ],
        "x-roles": [
          "ADMIN",
          "THU_THU"
        ]
      },
      "get": {
        "operationId": "MuonTra_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Trang, bắt đầu từ 1.",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Số dòng mỗi trang (1–100).",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "integer"
            }
          },
          {
            "name": "trangThai",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "DANG_MUON",
                "HOAN_TAT",
                "HUY"
              ]
            }
          },
          {
            "name": "maNguoiDung",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 20,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "total",
                    "page",
                    "limit"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PhieuMuonChiTietDto"
                      }
                    },
                    "total": {
                      "type": "integer",
                      "example": 42
                    },
                    "page": {
                      "type": "integer",
                      "example": 1
                    },
                    "limit": {
                      "type": "integer",
                      "example": 20
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dữ liệu gửi lên không hợp lệ (message là mảng lỗi validate) hoặc vi phạm CHECK của CSDL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Thiếu/sai/hết hạn JWT, hoặc tài khoản/người dùng đang bị khóa hay ngừng hoạt động",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Vai trò của tài khoản không được phép gọi endpoint này",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy bản ghi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "422": {
            "description": "Vi phạm quy tắc nghiệp vụ do procedure/trigger của CSDL báo (message là thông báo của CSDL)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Danh sách phiếu mượn (phân trang, lọc theo trạng thái/người mượn)",
        "tags": [
          "muon-tra"
        ],
        "x-roles": [
          "ADMIN",
          "THU_THU"
        ],
        "description": "**Quyền:** `ADMIN`, `THU_THU`"
      }
    },
    "/phieu-muon/{maPhieu}": {
      "get": {
        "description": "Cán bộ xem được mọi phiếu; bạn đọc chỉ xem phiếu của mình (phiếu của người khác: 403).\n\n**Quyền:** mọi tài khoản đã đăng nhập.",
        "operationId": "MuonTra_findOne",
        "parameters": [
          {
            "name": "maPhieu",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhieuMuonChiTietDto"
                }
              }
            }
          },
          "400": {
            "description": "Dữ liệu gửi lên không hợp lệ (message là mảng lỗi validate) hoặc vi phạm CHECK của CSDL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Thiếu/sai/hết hạn JWT, hoặc tài khoản/người dùng đang bị khóa hay ngừng hoạt động",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Vai trò của tài khoản không được phép gọi endpoint này",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy bản ghi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "422": {
            "description": "Vi phạm quy tắc nghiệp vụ do procedure/trigger của CSDL báo (message là thông báo của CSDL)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Chi tiết phiếu mượn",
        "tags": [
          "muon-tra"
        ]
      }
    },
    "/phieu-muon/{maPhieu}/huy": {
      "post": {
        "description": "Qua `sp_huy_phieu_muon`. Phiếu không tồn tại/đã đóng, hoặc đã có sách (hãy trả sách thay vì hủy) thì 422.\n\n**Quyền:** `ADMIN`, `THU_THU`",
        "operationId": "MuonTra_huy",
        "parameters": [
          {
            "name": "maPhieu",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhieuMuonChiTietDto"
                }
              }
            }
          },
          "400": {
            "description": "Dữ liệu gửi lên không hợp lệ (message là mảng lỗi validate) hoặc vi phạm CHECK của CSDL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Thiếu/sai/hết hạn JWT, hoặc tài khoản/người dùng đang bị khóa hay ngừng hoạt động",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Vai trò của tài khoản không được phép gọi endpoint này",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy bản ghi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "422": {
            "description": "Vi phạm quy tắc nghiệp vụ do procedure/trigger của CSDL báo (message là thông báo của CSDL)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Hủy phiếu mượn",
        "tags": [
          "muon-tra"
        ],
        "x-roles": [
          "ADMIN",
          "THU_THU"
        ]
      }
    },
    "/muon-tra/tra": {
      "post": {
        "description": "Qua `sp_tra_sach`. Trả quá hạn, hư hỏng hoặc mất thì DB tự lập phiếu phạt (xem `phieuPhats` trong kết quả). Bản sách không có lượt mượn đang mở thì 422; `tinhTrang` mặc định `BINH_THUONG`.\n\n**Quyền:** `ADMIN`, `THU_THU`",
        "operationId": "MuonTra_tra",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TraSachDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TraSachKetQuaDto"
                }
              }
            }
          },
          "400": {
            "description": "Dữ liệu gửi lên không hợp lệ (message là mảng lỗi validate) hoặc vi phạm CHECK của CSDL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Thiếu/sai/hết hạn JWT, hoặc tài khoản/người dùng đang bị khóa hay ngừng hoạt động",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Vai trò của tài khoản không được phép gọi endpoint này",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy bản ghi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "422": {
            "description": "Vi phạm quy tắc nghiệp vụ do procedure/trigger của CSDL báo (message là thông báo của CSDL)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Trả sách",
        "tags": [
          "muon-tra"
        ],
        "x-roles": [
          "ADMIN",
          "THU_THU"
        ]
      }
    },
    "/muon-tra/gia-han": {
      "post": {
        "description": "Cán bộ: `sp_gia_han` (mọi lượt mượn). Bạn đọc: `sp_gia_han_luot_muon` chỉ cho sách **của mình**; sách của người khác bị báo như không có lượt mượn (422, không phải 403). 422 khi: đã gia hạn tối đa số lần cho phép (tham số `SO_LAN_GIA_HAN_TOI_DA`), sách đã quá hạn, người dùng không hoạt động, hoặc có người đủ điều kiện đang đặt trước sách này.\n\n**Quyền:** mọi tài khoản đã đăng nhập.",
        "operationId": "MuonTra_giaHan",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GiaHanDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GiaHanKetQuaDto"
                }
              }
            }
          },
          "400": {
            "description": "Dữ liệu gửi lên không hợp lệ (message là mảng lỗi validate) hoặc vi phạm CHECK của CSDL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Thiếu/sai/hết hạn JWT, hoặc tài khoản/người dùng đang bị khóa hay ngừng hoạt động",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Vai trò của tài khoản không được phép gọi endpoint này",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy bản ghi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "422": {
            "description": "Vi phạm quy tắc nghiệp vụ do procedure/trigger của CSDL báo (message là thông báo của CSDL)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Gia hạn lượt mượn",
        "tags": [
          "muon-tra"
        ]
      }
    },
...
  },
  "info": {
    "title": "QLTV Nhóm 8 - API quản lý thư viện",
    "description": "API quản lý thư viện đại học (đồ án IE103, nhóm 8). Nghiệp vụ nằm trong MySQL (procedure, trigger, view); BE chỉ gọi lại và xác thực.\n\n**Cách dùng nhanh**\n1. `POST /auth/login` với tài khoản seed (ví dụ `ad001` / `AD001@Nhom8`), lấy `accessToken`.\n2. Bấm **Authorize** ở góc phải và dán token (không cần gõ chữ `Bearer`). Token được nhớ khi tải lại trang.\n3. Mỗi endpoint ghi rõ **Quyền** được gọi. Sai vai trò: 403. Token sai, hết hạn hoặc tài khoản bị khóa: 401.\n\n**Quy ước dữ liệu**\n- Khóa chính BIGINT (`id`) là **chuỗi**, ví dụ `\"13\"`. Các mã (`maSach`, `maBanSach`, `maNguoiDung`...) là chuỗi và là thứ dùng khi gọi API nghiệp vụ.\n- Endpoint danh sách có phân trang trả `{ data, total, page, limit }` (`page` từ 1, `limit` tối đa 100).\n- Cột ngày kiểu DATE trả dạng ISO `YYYY-MM-DDT00:00:00.000Z`.\n- Tiền (VND) từ báo cáo và `/me` là number; `soTien`/`giaBia` đọc từ bảng là chuỗi thập phân (Prisma Decimal).\n- Báo cáo (`/bao-cao/*`), `GET /sach` và `/me/*` giữ tên cột snake_case của view; các endpoint còn lại dùng camelCase.\n- Lỗi luôn có dạng `{ statusCode, message }`. 422 là lỗi nghiệp vụ do CSDL báo, `message` là thông báo cho người dùng (tiếng Việt không dấu).",
    "version": "2.2",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 422
          },
          "error": {
            "type": "string",
            "example": "Unprocessable Entity",
            "description": "Tên mã HTTP (chỉ có ở lỗi do bộ lọc CSDL và một số lỗi của Nest)"
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": "Nguoi dung dang no phat",
            "description": "Chuỗi thông báo; lỗi validate dữ liệu gửi lên (400) là mảng các thông báo"
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      },
...
      "TaoPhieuMuonDto": {
        "type": "object",
        "properties": {
          "maNguoiDung": {
            "type": "string",
            "maxLength": 20
          },
          "maBanSachs": {
            "description": "Tối đa 5 là mức mặc định (tham_so SO_SACH_TOI_DA); DB mới là nơi quyết định.",
            "uniqueItems": true,
            "minItems": 1,
            "maxItems": 20,
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 30
            }
          }
        },
        "required": [
          "maNguoiDung",
          "maBanSachs"
        ]
      },
      "TrangThaiPhieuMuon": {
        "type": "string",
        "enum": [
          "DANG_MUON",
          "HOAN_TAT",
          "HUY"
        ]
      },
      "NguoiDungTomTatDto": {
        "type": "object",
        "properties": {
          "maNguoiDung": {
            "type": "string",
            "example": "SV001"
          },
          "hoTen": {
            "type": "string",
            "example": "Võ Hoàng Nhiên"
          }
        },
        "required": [
          "maNguoiDung",
          "hoTen"
        ]
      },
      "TinhTrangTra": {
        "type": "string",
        "enum": [
          "BINH_THUONG",
          "HU_HONG",
          "MAT"
        ]
      },
...
      "LoaiPhat": {
        "type": "string",
        "enum": [
          "QUA_HAN",
          "HU_HONG",
          "MAT_SACH"
        ]
      },
...
      "PhieuPhatDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1",
            "description": "Mã phiếu phạt (dùng ở /phat/:id/...)",
            "nullable": false
          },
          "ctPhieuMuonId": {
            "type": "string",
            "example": "1",
            "nullable": false
          },
          "loaiPhat": {
            "nullable": false,
            "allOf": [
              {
                "$ref": "#/components/schemas/LoaiPhat"
              }
            ]
          },
          "soTien": {
            "type": "string",
            "example": "85000.00",
            "description": "Số tiền phạt (VND)",
            "nullable": false
          },
          "lyDo": {
            "type": "string",
            "nullable": true
          },
          "trangThai": {
            "nullable": false,
            "allOf": [
              {
                "$ref": "#/components/schemas/TrangThaiPhieuPhat"
              }
            ]
          },
          "ngayTao": {
            "type": "string",
            "format": "date-time"
          },
          "ngayThanhToan": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "ctPhieuMuonId",
          "loaiPhat",
          "soTien",
          "lyDo",
          "trangThai",
          "ngayTao",
          "ngayThanhToan"
        ]
      },
      "CtPhieuMuonChiTietDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1",
            "nullable": false
          },
          "phieuMuonId": {
            "type": "string",
            "example": "1",
            "nullable": false
          },
          "banSachId": {
            "type": "string",
            "example": "1",
            "nullable": false
          },
          "hanTra": {
            "type": "string",
            "format": "date-time",
            "description": "Hạn trả (kiểu DATE)"
          },
          "ngayTra": {
            "type": "string",
            "nullable": true,
            "format": "date-time",
            "description": "null = đang mượn"
          },
          "soLanGiaHan": {
            "type": "number",
            "description": "Số lần đã gia hạn"
          },
          "tinhTrangTra": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/TinhTrangTra"
              }
            ]
          },
          "banSachDangMuon": {
            "type": "string",
            "nullable": true,
            "example": "1",
            "description": "Cột sinh tự động (khác null khi đang mượn); không dùng ở FE"
          },
          "banSach": {
            "$ref": "#/components/schemas/MuonBanSachDto"
          },
          "phieuPhats": {
            "description": "Phiếu phạt (quá hạn/hư hỏng/mất) sinh ra từ lượt mượn này",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PhieuPhatDto"
            }
          }
        },
        "required": [
          "id",
          "phieuMuonId",
          "banSachId",
          "hanTra",
          "ngayTra",
          "soLanGiaHan",
          "tinhTrangTra",
          "banSachDangMuon",
          "banSach",
          "phieuPhats"
        ]
      },
      "PhieuMuonChiTietDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1",
            "nullable": false
          },
          "maPhieu": {
            "type": "string",
            "nullable": true,
            "example": "PM001"
          },
          "nguoiDungId": {
            "type": "string",
            "example": "1",
            "description": "Người mượn (nguoi_dung.id)",
            "nullable": false
          },
          "nhanVienId": {
            "type": "string",
            "example": "1",
            "description": "Cán bộ lập phiếu (nguoi_dung.id)",
            "nullable": false
          },
          "ngayMuon": {
            "type": "string",
            "format": "date-time",
            "description": "Ngày mượn (kiểu DATE)"
          },
          "trangThai": {
            "nullable": false,
            "allOf": [
              {
                "$ref": "#/components/schemas/TrangThaiPhieuMuon"
              }
            ]
          },
          "nguoiDung": {
            "$ref": "#/components/schemas/NguoiDungTomTatDto"
          },
          "nhanVien": {
            "$ref": "#/components/schemas/NguoiDungTomTatDto"
          },
          "ctPhieuMuons": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CtPhieuMuonChiTietDto"
            }
          }
        },
        "required": [
          "id",
          "maPhieu",
          "nguoiDungId",
          "nhanVienId",
          "ngayMuon",
          "trangThai",
          "nguoiDung",
          "nhanVien",
          "ctPhieuMuons"
        ]
      },
      "TraSachDto": {
        "type": "object",
        "properties": {
          "tinhTrang": {
            "description": "Tình trạng sách khi trả; HU_HONG/MAT sinh phiếu phạt. Mặc định BINH_THUONG.",
            "default": "BINH_THUONG",
            "allOf": [
              {
                "$ref": "#/components/schemas/TinhTrangTra"
              }
            ]
          },
          "maBanSach": {
            "type": "string",
            "maxLength": 30
          }
        },
        "required": [
          "maBanSach"
        ]
      },
      "MuonPhieuMaDto": {
        "type": "object",
        "properties": {
          "maPhieu": {
            "type": "string",
            "example": "PM001"
          }
        },
        "required": [
          "maPhieu"
        ]
      },
      "TraSachKetQuaDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1",
            "nullable": false
          },
          "phieuMuonId": {
            "type": "string",
            "example": "1",
            "nullable": false
          },
          "banSachId": {
            "type": "string",
            "example": "1",
            "nullable": false
          },
          "hanTra": {
            "type": "string",
            "format": "date-time",
            "description": "Hạn trả (kiểu DATE)"
          },
          "ngayTra": {
            "type": "string",
            "nullable": true,
            "format": "date-time",
            "description": "null = đang mượn"
          },
          "soLanGiaHan": {
            "type": "number",
            "description": "Số lần đã gia hạn"
          },
          "tinhTrangTra": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/TinhTrangTra"
              }
            ]
          },
          "banSachDangMuon": {
            "type": "string",
            "nullable": true,
            "example": "1",
            "description": "Cột sinh tự động (khác null khi đang mượn); không dùng ở FE"
          },
          "phieuPhats": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PhieuPhatDto"
            }
          },
          "phieuMuon": {
            "$ref": "#/components/schemas/MuonPhieuMaDto"
          }
        },
        "required": [
          "id",
          "phieuMuonId",
          "banSachId",
          "hanTra",
          "ngayTra",
          "soLanGiaHan",
          "tinhTrangTra",
          "banSachDangMuon",
          "phieuPhats",
          "phieuMuon"
        ]
      },
      "GiaHanDto": {
        "type": "object",
        "properties": {
          "maBanSach": {
            "type": "string",
            "maxLength": 30
          },
          "soNgay": {
            "type": "number",
            "minimum": 1,
            "maximum": 365
          }
        },
        "required": [
          "maBanSach",
          "soNgay"
        ]
      },
      "GiaHanKetQuaDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1",
            "nullable": false
          },
          "phieuMuonId": {
            "type": "string",
            "example": "1",
            "nullable": false
          },
          "banSachId": {
            "type": "string",
            "example": "1",
            "nullable": false
          },
          "hanTra": {
            "type": "string",
            "format": "date-time",
            "description": "Hạn trả (kiểu DATE)"
          },
          "ngayTra": {
            "type": "string",
            "nullable": true,
            "format": "date-time",
            "description": "null = đang mượn"
          },
          "soLanGiaHan": {
            "type": "number",
            "description": "Số lần đã gia hạn"
          },
          "tinhTrangTra": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/TinhTrangTra"
              }
            ]
          },
          "banSachDangMuon": {
            "type": "string",
            "nullable": true,
            "example": "1",
            "description": "Cột sinh tự động (khác null khi đang mượn); không dùng ở FE"
          },
          "phieuMuon": {
            "$ref": "#/components/schemas/MuonPhieuMaDto"
          }
        },
        "required": [
          "id",
          "phieuMuonId",
          "banSachId",
          "hanTra",
          "ngayTra",
          "soLanGiaHan",
          "tinhTrangTra",
          "banSachDangMuon",
          "phieuMuon"
        ]
      },
...
    }
  }
}